First steps towards generic interface for drivers
