+++ BSP/src/samv71/pio_driver.h
@@ -1,0 +1,4 @@
+#ifndef _PIO_DRIVER_H_
+#define _PIO_DRIVER_H_
+
+#endif // _PIO_DRIVER_H_
